+++ src/game/nlcore/elements/type.ts
@@ -1,0 +1,5 @@
+import {TagDefinition} from "@core/elements/image";
+
+export type {
+    TagDefinition,
+};